lint errors fixed
