+++ package.json
@@ -66,9 +66,9 @@
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@grafana/e2e": "8.4.1",
-    "@grafana/e2e-selectors": "8.4.1",
-    "@grafana/toolkit": "8.4.1",
+    "@grafana/e2e": "8.4.3",
+    "@grafana/e2e-selectors": "8.4.3",
+    "@grafana/toolkit": "8.4.3",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^13.1.9",
