+++ pkg/infinity/query.go
@@ -1,5 +1,16 @@
 package infinity
 
+const (
+	QueryTypeJSON    = "json"
+	QueryTypeCSV     = "csv"
+	QueryTypeTSV     = "tsv"
+	QueryTypeXML     = "xml"
+	QueryTypeGraphQL = "graphql"
+	QueryTypeHTML    = "html"
+	QueryTypeUQL     = "uql"
+	QueryTypeGROQ    = "groq"
+)
+
 type URLOptionKeyValuePair struct {
 	Key   string `json:"key"`
 	Value string `json:"value"`
@@ -13,7 +24,7 @@ type URLOptions struct {
 
 type Query struct {
 	RefID      string     `json:"refId"`
-	Type       string     `json:"type"`   // 'json' | 'csv' | 'tsv' | 'xml' | 'graphql' | 'html' | 'series' | 'global' | 'uql' | 'groq'
+	Type       string     `json:"type"`   // 'json' | 'csv' | 'tsv' | 'xml' | 'graphql' | 'html' | 'uql' | 'groq' | 'series' | 'global'
 	Format     string     `json:"format"` // 'table' | 'timeseries' | 'dataframe' | 'as-is' | 'node-graph-nodes' | 'node-graph-edges'
 	Source     string     `json:"source"` // 'url' | 'inline' | 'random-walk' | 'expression'
 	URL        string     `json:"url"`
